pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "qbit-arr"
version = "0.1.0"
description = "Media file relationship and orphan detection tool for qBittorrent, Radarr, and Sonarr"
authors = [{name = "Milan"}]
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "click>=8.1.7",
    "rich>=13.7.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "qbittorrent-api>=2024.1.59",
    "pyarr>=5.2.0",
    "requests>=2.31.0",
    "pyyaml>=6.0.1",
    "psutil>=5.9.8",
    "websockets>=12.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.12.0",
    "ruff>=0.1.9",
    "mypy>=1.8.0",
]

[project.scripts]
qbit-arr = "qbit_arr.cli.commands:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 100
target-version = ["py39"]

[tool.ruff]
line-length = 100
target-version = "py39"

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
